+++ HekaMOLD.Business/Models/DataTransfer/Production/KnitYarnModel.cs
@@ -18,7 +18,7 @@ namespace HekaMOLD.Business.Models.DataTransfer.Production
         public decimal? ReportWireCount { get; set; }
 
         //Metre Tel Sayisi
-        public decimal? MeterWireCount { get; set; }
+        public int? MeterWireCount { get; set; }
         public decimal? Gramaj { get; set; }
         public int? Density { get; set; }
